+++ db/models/orderProduct.js
@@ -2,7 +2,7 @@
 
 // bcrypt docs: https://www.npmjs.com/package/bcrypt
 const bcrypt = require('bcryptjs')
-    , {INTEGER, FLOAT} = require('sequelize')
+    , {INTEGER, DECIMAL} = require('sequelize')
 
 module.exports = db => db.define('orderProduct', {
   quantity: {
@@ -11,7 +11,7 @@ module.exports = db => db.define('orderProduct', {
     // defaultValue: 1
   },
   unitPrice: {
-    type: FLOAT(2)
+    type: DECIMAL(10,2)
     // we can add this from the front end. Otherwse:
     // set: this.setDataValue('unitPrice', this.product.price)
   }
